[Issue #27] Ctrl + C is now detected and calls the dispatcher to terminate the current process. Getting a bit confusing as to what is allowed to call what in terms of the OS and hardware, but I think this is correct where the dispatcher handles CPU interactions.

+++ distrib/host/dispatcher.js
@@ -9,9 +9,10 @@ var TSOS;
             _CPU.isExecuting = true;
         };
         Dispatcher.prototype.terminateCurrentProcess = function () {
-            if (_CPU.PCB) {
+            if (_CPU.PCB && _CPU.PCB.state != "terminated") {
                 _CPU.saveState();
                 _CPU.PCB.terminate();
+                _CPU.isExecuting = false;
             }
         };
         return Dispatcher;
